move theme in own file
